requirements.txt

fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]==1.7.4 
bcrypt==4.0.1

